docs(drawer): add drawer docs
